+++ Tests/Integration/MSFT_xDisk.Integration.Tests.ps1
@@ -28,11 +28,11 @@ try
         Return
     }
 
-    #region Integration Tests
-    $ConfigFile = Join-Path -Path $PSScriptRoot -ChildPath "$($script:DSCResourceName).config.ps1"
+    #region Integration Tests for DiskNumber
+    $ConfigFile = Join-Path -Path $PSScriptRoot -ChildPath "$($script:DSCResourceName).Number.config.ps1"
     . $ConfigFile -Verbose -ErrorAction Stop
 
-    Describe "$($script:DSCResourceName)_Integration" {
+    Describe "$($script:DSCResourceName)_DiskNumber_Integration" {
         Context 'Partition and format newly provisioned disk and assign a Drive Letter' {
             # Create a VHDx and attach it to the computer
             $VHDPath = Join-Path -Path $TestDrive `
@@ -63,7 +63,7 @@ try
                         )
                     }
 
-                    & "$($script:DSCResourceName)_Config" `
+                    & "$($script:DSCResourceName)_Number_Config" `
                         -OutputPath $TestDrive `
                         -ConfigurationData $ConfigData
                     Start-DscConfiguration -Path $TestDrive `
@@ -78,9 +78,72 @@ try
 
             It 'Should have set the resource and all the parameters should match' {
                 $current = Get-DscConfiguration | Where-Object {
-                    $_.ConfigurationName -eq "$($script:DSCResourceName)_Config"
+                    $_.ConfigurationName -eq "$($script:DSCResourceName)_Number_Config"
                 }
-                $current.DiskNumber       | Should Be $Disk.DiskNumber
+                $current.DiskNumber       | Should Be $Disk.Number
+                $current.DriveLetter      | Should Be $DriveLetter
+                $current.FSLabel          | Should Be $FSLabel
+            }
+
+            Dismount-DiskImage -ImagePath $VHDPath -StorageType VHDx
+            Remove-Item -Path $VHDPath -Force
+        }
+    }
+    #endregion
+
+    #region Integration Tests for DiskId
+    $ConfigFile = Join-Path -Path $PSScriptRoot -ChildPath "$($script:DSCResourceName).UniqueId.config.ps1"
+    . $ConfigFile -Verbose -ErrorAction Stop
+
+    Describe "$($script:DSCResourceName)_DiskUniqueId_Integration" {
+        Context 'Partition and format newly provisioned disk and assign a Drive Letter' {
+            # Create a VHDx and attach it to the computer
+            $VHDPath = Join-Path -Path $TestDrive `
+                -ChildPath 'TestDisk.vhdx'
+            New-VHD -Path $VHDPath -SizeBytes 1GB -Dynamic
+            Mount-DiskImage -ImagePath $VHDPath -StorageType VHDX -NoDriveLetter
+            $Disk = Get-Disk | Where-Object -FilterScript {
+                $_.Location -eq $VHDPath
+            }
+            $FSLabel = 'TestDisk'
+
+            # Get a spare drive letter
+            $LastDrive = ((Get-Volume).DriveLetter | Sort-Object | Select-Object -Last 1)
+            $DriveLetter = [char](([int][char]$LastDrive)+1)
+
+            #region DEFAULT TESTS
+            It 'Should compile without throwing' {
+                {
+                    # This is so that the
+                    $ConfigData = @{
+                        AllNodes = @(
+                            @{
+                                NodeName      = 'localhost'
+                                DriveLetter   = $DriveLetter
+                                DiskUniqueId  = $Disk.UniqueId
+                                FSLabel       = $FSLabel
+                            }
+                        )
+                    }
+
+                    & "$($script:DSCResourceName)_UniqueId_Config" `
+                        -OutputPath $TestDrive `
+                        -ConfigurationData $ConfigData
+                    Start-DscConfiguration -Path $TestDrive `
+                        -ComputerName localhost -Wait -Verbose -Force
+                } | Should not throw
+            }
+
+            It 'should be able to call Get-DscConfiguration without throwing' {
+                { Get-DscConfiguration -Verbose -ErrorAction Stop } | Should Not throw
+            }
+            #endregion
+
+            It 'Should have set the resource and all the parameters should match' {
+                $current = Get-DscConfiguration | Where-Object {
+                    $_.ConfigurationName -eq "$($script:DSCResourceName)_UniqueId_Config"
+                }
+                $current.DiskUniqueId     | Should Be $Disk.UniqueId
                 $current.DriveLetter      | Should Be $DriveLetter
                 $current.FSLabel          | Should Be $FSLabel
             }